Refactored dummy motor.
