+++ lys_instr/__init__.py
@@ -1,5 +1,4 @@
-from .MultiMotor import MultiMotorInterface
-from .MultiController import MultiControllerInterface, MultiSwitchInterface
+from .MultiController import MultiControllerInterface, MultiSwitchInterface, MultiMotorInterface
 from .MultiDetector import MultiDetectorInterface
 from .DataStorage import DataStorage
 from .PreCorrection import PreCorrector
